+++ naga/src/front/wgsl/parse/directive/enable_extension.rs
@@ -22,6 +22,7 @@ pub(crate) struct EnableExtensions {
     draw_index: bool,
     primitive_index: bool,
     per_vertex: bool,
+    wgpu_binding_array: bool,
 }
 
 impl EnableExtensions {
@@ -38,6 +39,7 @@ impl EnableExtensions {
             draw_index: false,
             primitive_index: false,
             per_vertex: false,
+            wgpu_binding_array: false,
         }
     }
 
@@ -59,6 +61,7 @@ impl EnableExtensions {
             ImplementedEnableExtension::DrawIndex => &mut self.draw_index,
             ImplementedEnableExtension::PrimitiveIndex => &mut self.primitive_index,
             ImplementedEnableExtension::PerVertex => &mut self.per_vertex,
+            ImplementedEnableExtension::WgpuBindingArray => &mut self.wgpu_binding_array,
         };
         *field = true;
     }
@@ -79,6 +82,7 @@ impl EnableExtensions {
             ImplementedEnableExtension::DrawIndex => self.draw_index,
             ImplementedEnableExtension::PrimitiveIndex => self.primitive_index,
             ImplementedEnableExtension::PerVertex => self.per_vertex,
+            ImplementedEnableExtension::WgpuBindingArray => self.wgpu_binding_array,
         }
     }
 
@@ -132,6 +136,7 @@ impl EnableExtension {
     const PRIMITIVE_INDEX: &'static str = "primitive_index";
     const DRAW_INDEX: &'static str = "draw_index";
     const PER_VERTEX: &'static str = "wgpu_per_vertex";
+    const BINDING_ARRAY: &'static str = "wgpu_binding_array";
 
     /// Convert from a sentinel word in WGSL into its associated [`EnableExtension`], if possible.
     pub(crate) fn from_ident(word: &str, span: Span) -> Result<'_, Self> {
@@ -156,6 +161,7 @@ impl EnableExtension {
             Self::DRAW_INDEX => Self::Implemented(ImplementedEnableExtension::DrawIndex),
             Self::PRIMITIVE_INDEX => Self::Implemented(ImplementedEnableExtension::PrimitiveIndex),
             Self::PER_VERTEX => Self::Implemented(ImplementedEnableExtension::PerVertex),
+            Self::BINDING_ARRAY => Self::Implemented(ImplementedEnableExtension::WgpuBindingArray),
             _ => return Err(Box::new(Error::UnknownEnableExtension(span, word))),
         })
     }
@@ -177,6 +183,7 @@ impl EnableExtension {
                 ImplementedEnableExtension::PrimitiveIndex => Self::PRIMITIVE_INDEX,
                 ImplementedEnableExtension::WgpuRayTracingPipeline => Self::RAY_TRACING_PIPELINE,
                 ImplementedEnableExtension::PerVertex => Self::PER_VERTEX,
+                ImplementedEnableExtension::WgpuBindingArray => Self::BINDING_ARRAY,
             },
             Self::Unimplemented(kind) => match kind {
                 UnimplementedEnableExtension::Subgroups => Self::SUBGROUPS,
@@ -227,6 +234,8 @@ pub enum ImplementedEnableExtension {
     PrimitiveIndex,
     /// Enables the `wgpu_per_vertex` extension, allows using `@interpolate(per_vertex)` attribute in WGSL, native only.
     PerVertex,
+    /// Enables the `wgpu_binding_array` extension, native only.
+    WgpuBindingArray,
 }
 
 impl ImplementedEnableExtension {
@@ -243,6 +252,7 @@ impl ImplementedEnableExtension {
         Self::DrawIndex,
         Self::PrimitiveIndex,
         Self::PerVertex,
+        Self::WgpuBindingArray,
     ];
 
     /// Returns slice of all variants of [`ImplementedEnableExtension`].
@@ -265,6 +275,15 @@ impl ImplementedEnableExtension {
             Self::DrawIndex => C::DRAW_INDEX,
             Self::PrimitiveIndex => C::PRIMITIVE_INDEX,
             Self::PerVertex => C::PER_VERTEX,
+            Self::WgpuBindingArray => C::BUFFER_BINDING_ARRAY
+                .union(C::BUFFER_BINDING_ARRAY_NON_UNIFORM_INDEXING)
+                .union(C::STORAGE_BUFFER_BINDING_ARRAY)
+                .union(C::STORAGE_BUFFER_BINDING_ARRAY_NON_UNIFORM_INDEXING)
+                .union(C::STORAGE_TEXTURE_BINDING_ARRAY)
+                .union(C::STORAGE_TEXTURE_BINDING_ARRAY_NON_UNIFORM_INDEXING)
+                .union(C::TEXTURE_AND_SAMPLER_BINDING_ARRAY)
+                .union(C::TEXTURE_AND_SAMPLER_BINDING_ARRAY_NON_UNIFORM_INDEXING)
+                .union(C::ACCELERATION_STRUCTURE_BINDING_ARRAY),
         }
     }
 }